Tried fixing create task member selection problem (i failed)

+++ app/src/main/java/com/example/ProjectManager/activities/AddMemberActivity.java
@@ -144,7 +144,7 @@ public class AddMemberActivity extends AppCompatActivity {
                     if (users != null) {
                         for (UserResponseDto user : users) {
                             Member member = new Member(
-                                    user.getId().intValue(),
+                                    user.getId(),
                                     user.getFirstName() + " " + user.getLastName(),
                                     "" // No role from API response
                             );
